+++ resources/js/notifier.js
@@ -20,15 +20,14 @@
   const path = require('path')
   const url = require('url')
   const https = require('https')
-  const HttpsProxyAgent = require('https-proxy-agent')
   const storage = require('electron-json-storage')
   const encrypt = require('./decrypt.js')
   const preferences = require('./preferences.js')
+  const preparer = require('./websocket-preparer.js')
 
   let lastNotificationTime = new Date().getTime()
   let currentNotification = null
   let windowProvider = null
-  let proxyAgent = null
 
   var notify = (title, snippet, conversation_id, provider) => {
     windowProvider = provider
@@ -147,7 +146,7 @@
           'Content-Type': 'application/json',
           'Content-Length': Buffer.byteLength(data)
         },
-        agent: getProxyAgent()
+        agent: preparer.getProxyAgent()
       }
 
       var req = https.request(options)
@@ -176,7 +175,7 @@
           'Content-Type': 'application/json',
           'Content-Length': Buffer.byteLength(data)
         },
-        agent: getProxyAgent()
+        agent: preparer.getProxyAgent()
       }
 
       var req = https.request(options)
@@ -204,7 +203,7 @@
           'Content-Type': 'application/json',
           'Content-Length': Buffer.byteLength(data)
         },
-        agent: getProxyAgent()
+        agent: preparer.getProxyAgent()
       }
 
       var req = https.request(options)
@@ -226,17 +225,5 @@
     return getRandomInt(1, 922337203685477)
   }
 
-  function getProxyAgent() {
-    // use the PULSE_PROXY or HTTPS_PROXY environment variable to determine if we should use a proxy
-    var envProxy = process.env.PULSE_PROXY || process.env.HTTPS_PROXY
-    if (!proxyAgent && envProxy) {
-      console.log("Attempting to get a proxy agent with \"" + envProxy + "\"")
-      proxyAgent = HttpsProxyAgent(envProxy)
-    }
-
-    return proxyAgent
-  }
-
   module.exports.notify = notify
-  module.exports.getProxyAgent = getProxyAgent
 }())
